+++ prince-archiver/prince_archiver/entrypoints/worker/worker.py
@@ -17,8 +17,8 @@ from prince_archiver.config import ArchiveWorkerSettings
 from prince_archiver.log import configure_logging
 from prince_archiver.models import init_mappers
 from prince_archiver.service_layer.handlers.archive import add_data_archive_entry
-from prince_archiver.service_layer.handlers.exporter import persist_imaging_event_export
-from prince_archiver.service_layer.handlers.importer import import_imaging_event
+from prince_archiver.service_layer.handlers.export import persist_imaging_event_export
+from prince_archiver.service_layer.handlers.ingest import import_imaging_event
 from prince_archiver.service_layer.messagebus import MessageBus
 from prince_archiver.service_layer.messages import (
     AddDataArchiveEntry,
